move strip_computed_fields and _collect_computed_field_names
